First scenario now working

+++ acme_challenge_dispatcher_test/features/steps/step_definitions.py
@@ -1,6 +1,7 @@
 import json
 import logging
 import os
+import requests
 import subprocess
 from time import sleep, time
 
@@ -37,35 +38,35 @@ def check_namespace_exists(context, namespace):
 def deploy_acme_solver_pod(context):
     script_dir = os.path.dirname(__file__)
     full_file_path = script_dir + "/test-data/acme-solver-pod-template.yaml"
-    with open(full_file_path) as f:
-         pod_template = yaml.safe_load(f)
 
     table = context.table
     name = table[0]['name']
     port = int(table[0]['port'])
     token = table[0]['token']
     key = table[0]['key']
+    domain = table[0]['domain']
 
-    delete_pod_if_exists(context.api_client, table[0]['name'], context.namespace)
+    if not does_pod_exist(context.api_client, name, context.namespace):
+        with open(full_file_path) as f:
+            pod_template = yaml.safe_load(f)
 
-    # Update the pod name
-    pod_template['metadata']['name'] = name
+        pod_template['metadata']['name'] = name
 
-    # Update the container args and port
-    container = pod_template['spec']['containers'][0]
-    container['args'] = [
-        f'--listen-port={port}',
-        '--domain=example.com',
-        f'--token={token}',
-        f'--key={key}'
-    ]
-    container['ports'][0]['containerPort'] = port
+        # Update the container args and port
+        container = pod_template['spec']['containers'][0]
+        container['args'] = [
+            f'--listen-port={port}',
+            f'--domain={domain}',
+            f'--token={token}',
+            f'--key={key}'
+        ]
+        container['ports'][0]['containerPort'] = port
 
-    v1 = context.api_client
-    delete_pod_if_exists(v1, name, context.namespace)
-    v1.create_namespaced_pod(namespace=context.namespace, body=pod_template)
+        v1 = context.api_client
+        delete_pod_if_exists(v1, name, context.namespace)
+        v1.create_namespaced_pod(namespace=context.namespace, body=pod_template)
 
-    is_pod_running_and_ready(v1, context.namespace, name)
+        is_pod_running_and_ready(v1, context.namespace, name)
 
 @step('I forward the port {pod_port:d} of the pod "{pod_name}" to port {host_port:d}')
 def forward_port(context, pod_port, pod_name, host_port):
@@ -91,9 +92,9 @@ def forward_port(context, pod_port, pod_name, host_port):
         print(f"Failed to forward port: {e}")
 
 
-@step('I wait for 1 minute')
-def wait_for_one_minute(context):
-    sleep(60)
+@step('I wait for {seconds:d} seconds')
+def wait_for_seconds(context, seconds):
+    sleep(seconds)
 
 @step('the service-account "{service_account_name}" exists')
 def step_impl(context, service_account_name):
@@ -118,22 +119,21 @@ def deploy_acme_challenge_dispatcher_pod(context):
     namespace = context.namespace
     table = context.table
 
-    # Read the pod template
-    pod_template_path = os.path.join(os.path.dirname(__file__), 'test-data', 'acme-challenge-dispatcher-pod-template.yaml')
-    with open(pod_template_path) as f:
-        pod_template = yaml.safe_load(f)
+    if not does_pod_exist(context.api_client, table[0]['name'], namespace):
 
-    pod_template['metadata']['name'] = table[0]['name']
-    pod_template['spec']['containers'][0]['image'] = table[0]['image']
+        pod_template_path = os.path.join(os.path.dirname(__file__), 'test-data', 'acme-challenge-dispatcher-pod-template.yaml')
+        with open(pod_template_path) as f:
+            pod_template = yaml.safe_load(f)
 
-    delete_pod_if_exists(context.api_client, table[0]['name'], namespace)
+        pod_template['metadata']['name'] = table[0]['name']
+        pod_template['spec']['containers'][0]['image'] = table[0]['image']
 
-    # Create the pod in the specified namespace
-    v1 = context.api_client
-    v1.create_namespaced_pod(namespace=namespace, body=pod_template)
-    print(f"Pod '{table[0]['name']}' with image '{table[0]['image']}' created in namespace '{namespace}'")
+        # Create the pod in the specified namespace
+        v1 = context.api_client
+        v1.create_namespaced_pod(namespace=namespace, body=pod_template)
+        print(f"Pod '{table[0]['name']}' with image '{table[0]['image']}' created in namespace '{namespace}'")
 
-    is_pod_running_and_ready(v1, namespace, pod_template['metadata']['name'])
+        is_pod_running_and_ready(v1, namespace, pod_template['metadata']['name'])
 
 @step('I stop forwarding the port {pod_port:d} of the pod "{pod_name}"')
 def stop_forwarding_port(context, pod_port, pod_name):
@@ -153,6 +153,19 @@ def stop_forwarding_port(context, pod_port, pod_name):
     else:
         print(f"No port forwarding process found for pod '{pod_name}' on port {pod_port}")
 
+@step('I do a GET request to {port:d} with the following parameters')
+def do_get_request(context, port):
+    params = context.table[0]
+    url = f"http://localhost:{port}{params['url']}"
+    headers = {'Host': params['host']}
+
+    response = requests.get(url, headers=headers)
+    context.response = response
+
+@step('the response should be "{expected_response}"')
+def check_response(context, expected_response):
+    assert context.response.text == expected_response, f"Expected '{expected_response}', but got '{context.response.text}'"
+
 def delete_pod(v1, pod_name, namespace):
     return v1.delete_namespaced_pod(name=pod_name, namespace=namespace)
 
@@ -168,6 +181,16 @@ def delete_pod_if_exists(v1, pod_name, namespace):
             has_deleted = json.loads(e.body)['code'] == 404
             if has_deleted:
                 return
+
+def does_pod_exist(v1, pod_name, namespace):
+    try:
+        v1.read_namespaced_pod(name=pod_name, namespace=namespace)
+        return True
+    except client.exceptions.ApiException as e:
+        if e.status == 404:
+            return False
+        else:
+            raise
 
 
 def is_pod_running_and_ready(v1, namespace, pod_name, timeout=60):
